requirements.txt
pandas
requests
python-dotenv
yagmail
openpyxl



